% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfilePlotUI.R
\name{createSubjectProfileFromShinyInput}
\alias{createSubjectProfileFromShinyInput}
\title{Create subject profile plot(s) based on parameters from the Shiny UI}
\usage{
createSubjectProfileFromShinyInput(input, results)
}
\arguments{
\item{input}{Shiny input object.}

\item{results}{reactiveValues object with list of results created from the server script.}
}
\value{
subject profile plot(s)
}
\description{
Create subject profile plot(s) based on parameters from the Shiny UI
}
\author{
Laure Cougnaud
}
