% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfilePlotUI.R
\name{createSubjectProfileType}
\alias{createSubjectProfileType}
\title{Create subject profile plot of specified type}
\usage{
createSubjectProfileType(listParams, type)
}
\arguments{
\item{listParams}{list of parameters}

\item{type}{string with type of profile plot}
}
\value{
subject profile plot(s)
}
\description{
Create subject profile plot of specified type
}
\author{
Laure Cougnaud
}
