% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultModules.R
\name{getDefaultModules}
\alias{getDefaultModules}
\title{Get default modules}
\usage{
getDefaultModules(datasetNames)
}
\arguments{
\item{datasetNames}{names of the dataset}
}
\value{
character vector with default modules
}
\description{
Get default modules
}
\author{
Laure Cougnaud
}
