% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUIParamModule.R
\name{getModuleParams}
\alias{getModuleParams}
\title{Get list of module parameters.}
\usage{
getModuleParams(input, use = c("UI", "plotFunction"), results = NULL)
}
\arguments{
\item{input}{Shiny input object.}

\item{use}{String with use or the output, either:
\itemize{
\item{'UI': }{output will be displayed in the Shiny UI}
\item{'plotFunction': }{output will be used as input 
parameters for the plotting functions.
In this case, the \code{results} parameter should be specified.}
}}

\item{results}{reactiveValues object with list of results created from the server script.}
}
\value{
list of parameters
}
\description{
Get list of module parameters.
}
\author{
Laure Cougnaud
}
