% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getTimeVars}
\alias{getTimeVars}
\title{Get potential time variables in the dataset}
\usage{
getTimeVars(data)
}
\arguments{
\item{data}{data.frame with data}
}
\value{
character vector with potential time variables
}
\description{
Get potential time variables in the dataset
}
\author{
Laure Cougnaud
}
