% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUIParamModule.R
\name{getUIParamModule}
\alias{getUIParamModule}
\title{Get list of module parameters specified in the UI
(to enable display of previously saved module)}
\usage{
getUIParamModule(input)
}
\arguments{
\item{input}{Shiny input object}
}
\value{
list of parameters
}
\description{
Get list of module parameters specified in the UI
(to enable display of previously saved module)
}
\author{
Laure Cougnaud
}
