% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getVarLabelsForUI}
\alias{getVarLabelsForUI}
\title{Get variable labels for the interface.
This concatenate the variable code (column in data) and variable labels}
\usage{
getVarLabelsForUI(data, labelVars)
}
\arguments{
\item{data}{data.frame with data}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
named character vector with variables
}
\description{
Get variable labels for the interface.
This concatenate the variable code (column in data) and variable labels
}
\author{
Laure Cougnaud
}
