% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPatientProfilesVisShiny.R
\name{runPatientProfilesVisShiny}
\alias{runPatientProfilesVisShiny}
\title{Run the shiny application}
\usage{
runPatientProfilesVisShiny(installDependencies = FALSE, ...)
}
\arguments{
\item{installDependencies}{boolean, whether to first install packages listed
in the Suggests field of DESCRIPTION; default value is FALSE}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value
}
\description{
Run the shiny application
}
