% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{simpleCap}
\alias{simpleCap}
\title{capitalize the first letter of a word, from the help of the 'toupper' function}
\usage{
simpleCap(x, onlyFirst = FALSE, rev = FALSE)
}
\arguments{
\item{x}{string}

\item{onlyFirst}{logical, if TRUE (FALSE by default)
capitalize the first letter of the first forward only}

\item{rev}{logical, if TRUE (FALSE by default), set first letter to lower case (otherwise upper case)}
}
\value{
string with first letter capitalized
}
\description{
capitalize the first letter of a word, from the help of the 'toupper' function
}
\author{
author of the 'toupper' function?
}
