% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultModules.R
\name{getDefaultModules}
\alias{getDefaultModules}
\title{Get default modules a specific data.
This is based on the ADaM dataset names.}
\usage{
getDefaultModules(data)
}
\arguments{
\item{data}{list of ADaM datasets}
}
\value{
list of list of input parameter for modules
}
\description{
Get default modules a specific data.
This is based on the ADaM dataset names.
}
\author{
Laure Cougnaud
}
