% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getTimeVars}
\alias{getTimeVars}
\title{Get potential time variables in the dataset}
\usage{
getTimeVars(data, labelVars)
}
\arguments{
\item{data}{data.frame with data}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
named character vector with potential time variables
}
\description{
Get potential time variables in the dataset
}
\author{
Laure Cougnaud
}
